% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_overrides.R
\name{apply.initial.overrides}
\alias{apply.initial.overrides}
\title{Apply manual reformatting}
\usage{
apply.initial.overrides(data, filename)
}
\arguments{
\item{data}{input character vector of data to be processed}

\item{filename}{name of file containing user-specified string replacements}
}
\value{
character vector of the same length as input, with replacements applied
}
\description{
This function accepts user-submitted input->output pairs
and uses pattern matching via \code{gsub} to convert instances of
'input' to 'output' in all strings.
}
