% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_overrides.R
\name{apply.posthoc.overrides}
\alias{apply.posthoc.overrides}
\title{Apply arbitrary replacements overriding the results of NLP}
\usage{
apply.posthoc.overrides(raw.input, title.data, author.data, filename)
}
\arguments{
\item{raw.input}{input character vector of initial submissions}

\item{title.data}{input character vector containing post-NLP assigned titles}

\item{author.data}{input character vector containing post-NLP assigned authors}

\item{filename}{input character vector containing filename of replacements}
}
\value{
list with 'title' and 'author' entries with replacements applied
}
\description{
Especially when the total number of submissions is low or the number of unique
submissions is high, some things are just going to be incorrect. This method
allows arbitrary replacements to override the NLP corrections to the input.
The replacement file should be of format
}
\details{
full.initial.entry	title	author	replacement.title	replacement.author

Only fill in one of the three first columns; this will be the value that is
matched for replacement. Depending on which of the title and author (or both)
are specified, those same values will be replaced in the output.
}
