% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_overrides.R
\name{apply.standard.case}
\alias{apply.standard.case}
\title{Apply modified title case to certain entries}
\usage{
apply.standard.case(i, apply.to.all)
}
\arguments{
\item{i}{character vector of input data to possibly be adjusted}

\item{apply.to.all}{should the values be corrected regardless of whether they're
uniformly caps locked?}
}
\value{
the input data 'i' with requested adjustments applied
}
\description{
An override: some single submissions are CAPS LOCK ONLY which is just silly.
find them, and if they're literally all caps, title case them.
note the enforcement of the entire thing being caps lock: this is because
there are some author entries, in particular people who use pairs of initials
(e.g. AB Jones) that should not be rendered "Ab Jones" :(
}
