% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_known_matches.R
\name{check.known.matches}
\alias{check.known.matches}
\title{Take a file of known matches and use it to QC the
proposed results}
\usage{
check.known.matches(df, filename)
}
\arguments{
\item{df}{data.frame, final results from a run of \code{parse_raw_data}}

\item{filename}{character vector, file of known matches between input and output, or NA}
}
\description{
If the user has a curated set of results, specifically the
results of the initial training pass, this function will take
those curated data and use them to conduct regression testing,
more or less. This is complicated by the lack of deterministic
behavior in this package. Nevertheless, certainly any regression
is worth flagging, even if it's so that the relevant content can
be added to the postprocessing overrides.
}
