% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_handling.R
\name{handle.duplicates}
\alias{handle.duplicates}
\title{Remove rows with identical entries}
\usage{
handle.duplicates(df, output.prefix)
}
\arguments{
\item{df}{data.frame, raw data representation with no processing}

\item{output.prefix}{character vector, prefix of output logging file}
}
\value{
data.frame, input df with duplicate rows removed (leaving first instance)
}
\description{
There are some example instances of duplicated rows with
mysteriously identical entries, including strange replicated
typographical idiosyncracies. These can optionally be removed,
and their removal is reported to a logging file.
}
