% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_handling.R
\name{process.output}
\alias{process.output}
\title{Take data and format it for user}
\usage{
process.output(data, output.prefix, create.summary.data = TRUE)
}
\arguments{
\item{data}{either a data.frame of data from \code{parse_raw_data}, or the name of a file
generated by a run of \code{parse_raw_data}}

\item{output.prefix}{character vector, prefix for any output files generated by the function}

\item{create.summary.data}{logical, whether to provide a simplified summary result for
reporting final data by category}
}
\description{
This function either takes the results of a primary run of \code{parse_raw_data},
or the name of a file containing data from that run, and processes it. The idea
is that it either provides a report of all processing steps for user intervention,
or it takes the results of user intervention and reports a summary that can be used
for other things, like for example actual downstream applications, who'd have guessed it.
}
