% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_raw_data.R
\name{parse_raw_data}
\alias{parse_raw_data}
\title{Process raw submitter data}
\usage{
parse_raw_data(
  input.filename,
  output.prefix,
  preprocessing.overrides = NA,
  postprocessing.overrides = NA,
  known.matches = NA,
  remove.duplicates = TRUE,
  disable.nlp.correction = FALSE
)
}
\arguments{
\item{input.filename}{character string, file containing input string data}

\item{output.prefix}{character string, prefix of files that will contain output for user intervention}

\item{preprocessing.overrides}{character string, a file of Content/Replacement pairs delimited by TABs, or NA}

\item{postprocessing.overrides}{character string, a file of Content/Replacement pairs delimited by TABs, or NA}

\item{known.matches}{character string, a file of curated {Entry -> {Title, Author}} pairs, or NA}

\item{remove.duplicates}{logical, whether to remove rows in input data that are absolutely identical (will report on this if so)}

\item{disable.nlp.correction}{logical, whether to just report the data as detected from input, without applying NLP corrections}
}
\description{
This function loads an Excel spreadsheet containing formatted output
from an initial user submission, and parses it into a format more readily
usable by experts for manual inspection and correction. Some degree of
ML/NLP is applied to try to harmonize data across multiple (very) similiar
entries. This functionality can be disabled if desired.
}
